Project building with VS2014
Adding the SDL .lib and .dll files, built with VS2014 CTP
Headers will be read from the existing SDL install
transformer_type in Hexagrid now uses double

+++ util/Hexagrid.hpp
@@ -37,7 +37,7 @@ public:
 	using box_type = bgm::box<hex_point_type>;
 
 	template<std::size_t dim = 2, bool SameScale = true>
-	using transformer_type = trans::map_transformer<int, dim,dim,true, SameScale>;
+	using transformer_type = trans::map_transformer<double, dim,dim,true, SameScale>;
 
 
 private:
